循環参照が発生するのでdeserializer.tsに移動
